refactor: add compiler api to root
